Decouple modules from infrastructure
