+++ src/OrchardCore.Modules/OrchardCore.DynamicCache/TagHelpers/DynamicCacheTagHelper.cs
@@ -1,14 +1,11 @@
 using System;
-using System.IO;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
-using AngleSharp.Text;
 using Microsoft.AspNetCore.Html;
 using Microsoft.AspNetCore.Mvc.TagHelpers.Cache;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 using Microsoft.Extensions.Options;
 using OrchardCore.Abstractions.Pooling;
-using OrchardCore.DisplayManagement;
 using OrchardCore.Environment.Cache;
 
 namespace OrchardCore.DynamicCache.TagHelpers
