+++ src/OrchardCore.Modules/OrchardCore.Users/ViewModels/ResetPasswordViewModel.cs
@@ -2,7 +2,6 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Localization;
-using MimeKit;
 using OrchardCore.Email;
 
 namespace OrchardCore.Users.ViewModels
