+++ src/OrchardCore/OrchardCore.Infrastructure.Abstractions/Commands/CommandNameAttribute.cs
