+++ src/OrchardCore/OrchardCore.Infrastructure.Abstractions/Commands/OrchardSwitchesAttribute.cs
