+++ src/OrchardCore/OrchardCore.Infrastructure.Abstractions/Security/ApiAuthorizationOptions.cs
@@ -1,0 +1,9 @@
+using Microsoft.AspNetCore.Authentication;
+
+namespace OrchardCore.Security
+{
+    public class ApiAuthorizationOptions : AuthenticationSchemeOptions
+    {
+        public string ApiAuthenticationScheme { get; set; } = "Bearer";
+    }
+}
