+++ src/OrchardCore/OrchardCore.Setup.Core/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
 using Microsoft.Extensions.DependencyInjection;
-using OrchardCore.Entities;
 using OrchardCore.Setup.Services;
 
 namespace OrchardCore.Setup
@@ -16,7 +15,6 @@ namespace OrchardCore.Setup
         {
             services.AddScoped<ISetupService, SetupService>();
 
-            services.AddIdGeneration();
             services.AddSingleton<ISetupUserIdGenerator, SetupUserIdGenerator>();
 
             return services;
